requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.2
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
elevenlabs>=0.2.26
beautifulsoup4>=4.12.2
Pillow>=10.1.0
emoji>=2.8.0
pyperclip>=1.11.0